refactored Comment and Note into classes
Created a Note class which is a webview panel
Changed both Comment (decoration) and Note to be classes
